chore(build): fix versioning and uploading
We previously were uploading the contents of firmwarebin/ to s3 which as it
turns out is a checked-in file that never changes. That is not what we want to
do.

Instead,
- Remove firmwarebin/ as a checked-in file
- Instead, use it as a directory containing the properly named build output
- Copy built hex files into firmwarebin using the same value as the version
string embedded in the build itself
- Make that version string generation actually work properly when running on a
checked out tag (which is a detached HEAD, which git show-ref doesn't like;
instead, there's a bit of logic to check if we have a branch and if not use git
describe)
- Upload tha actual fresh build to s3 instead
- Also, on tag, update the github release, since without the firmwarebin/
directory checked in it's nice to have somewhere to point to for stable builds
- Because we need to use git symbolic-ref and git describe, check out the
TRAVIS_BRANCH.