refactor(simpleshell,kernel): move rev bits to own module
Instead of having the revision bits be parsed by the simpleshell, do it
in their own module that is exported by the kernel. This is accessible
with THEKERNEL->revision->pcb_revision and is an enum defined in the new
libs/Revision.h. PCB revision bits are available during or after
on_module_load (at least if the module isn't one directly built by the
kernel; if it is, you may need to check initialization order).