feat(currentcontrol): change current based on revision (#11)
* refactor(simpleshell,kernel): move rev bits to own module

Instead of having the revision bits be parsed by the simpleshell, do it
in their own module that is exported by the kernel. This is accessible
with THEKERNEL->revision->pcb_revision and is an enum defined in the new
libs/Revision.h. PCB revision bits are available during or after
on_module_load (at least if the module isn't one directly built by the
kernel; if it is, you may need to check initialization order).

* feat(currentcontrol): change digipot factor based on rev

Rev A and on motor controllers use a different digipot factor than
rev12(.5) and previous. That means we
- add a new config setting for rev12 and previous digipots
- parse the new config setting alongside the standard one
- check the revision bits in on_module_load and load the right one