refactor(endstops): Load endstop pins by revision
New revisions of Opentrons smoothie boards have different pin
configuration requirements. Among other things, the endstop circuits are
changing and require different pull settings to work properly. To keep
one firmware image that works on all smoothie board variants, we need to
specify different pin control settings for different revisions.

This commit allows specifying configurations for specific revisions that
are used by any revision higher than that (C>B>A>12>unknown). The
software will use the configuration for the highest revision at or under
the current revision (i.e. if the revision is B and there's a C and A
revision, it will use A).